Compile error

+++ database/migrations/2021_03_08_085735_create_invoice_processes_table.php
@@ -16,7 +16,7 @@ class CreateInvoiceProcessesTable extends Migration
         Schema::create('invoice_processes', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('invoice_id');
-            $table->integer('owner_id');
+            $table->unsignedBigInteger('owner_id');
             $table->tinyInteger('owner_role_id');
             $table->boolean('approved')->nullable()->default(null);
             $table->text('reason')->nullable()->default(null);
